+++ presto-main/src/main/java/com/facebook/presto/operator/SharedLookupSource.java
@@ -56,7 +56,7 @@ public final class SharedLookupSource
     }
 
     @Override
-    public long getJoinPosition(int position, Page page, int rawHash)
+    public long getJoinPosition(int position, Page page, long rawHash)
     {
         return lookupSource.getJoinPosition(position, page, rawHash);
     }
